+++ src/app/app.component.ts
@@ -4,7 +4,7 @@ import { Platform } from '@ionic/angular';
 import { SplashScreen } from '@ionic-native/splash-screen/ngx';
 import { StatusBar } from '@ionic-native/status-bar/ngx';
 import { DataSharingService } from '../services/data-sharing.service';
-import { UserServices } from 'src/services/userServices';
+import { UserServices } from 'src/services/user.services';
 
 @Component({
   selector: 'app-root',
